Solution to Check Leap Year

+++ Check Leap Year@@6788dd63490637cfc44e1b90/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 void main()
 {
-    int m;
+    long m;
     scanf("%d",&m);
     if ((m%4==0) && (m%400==0))
         printf("Not a Leap Year");
